Removing initial PoC code that separated the current cart and all unsubmitted orders stores.
