--- Website/src/rendering/src/hooks/useOcUnsubmittedOrders.ts
@@ -1,7 +1,0 @@
-import { OcUnsubmittedOrdersState } from '../redux/ocUnsubmittedOrders';
-import { useAppSelector } from '../redux/store';
-
-const useOcUnsubmittedOrders = (): OcUnsubmittedOrdersState =>
-  useAppSelector((s) => s.ocUnsubmittedOrders);
-
-export default useOcUnsubmittedOrders;
