+++ Website/src/rendering/src/redux/ocProvider.tsx
@@ -1,8 +1,7 @@
 import React, { FunctionComponent, useEffect, useState } from 'react';
 import { initializeAuth } from './ocAuth';
 import logout from './ocAuth/logout';
-import { retrieveOrders } from './ocUnsubmittedOrders';
-import { retrieveCart } from './ocCurrentCart';
+import { retrieveCart, retrieveOrders } from './ocCurrentCart';
 import { useAppDispatch, useAppSelector } from './store';
 import { getUser } from './ocUser';
 import { Configuration, Tokens } from 'ordercloud-javascript-sdk';
@@ -56,11 +55,10 @@ const OcProvider: FunctionComponent = ({ children }) => {
     setHasCheckedForToken(true);
   }, [dispatch, router]);
 
-  const { ocAuth, ocUser, ocCurrentCart, ocUnsubmittedOrders } = useAppSelector((s) => ({
+  const { ocAuth, ocUser, ocCurrentCart } = useAppSelector((s) => ({
     ocAuth: s.ocAuth,
     ocUser: s.ocUser,
     ocCurrentCart: s.ocCurrentCart,
-    ocUnsubmittedOrders: s.ocUnsubmittedOrders,
   }));
 
   useEffect(() => {
@@ -75,13 +73,11 @@ const OcProvider: FunctionComponent = ({ children }) => {
         }
         if (!ocCurrentCart.initialized) {
           dispatch(retrieveCart(null));
-        }
-        if (!ocUnsubmittedOrders.initialized) {
           dispatch(retrieveOrders());
         }
       }
     }
-  }, [dispatch, ocAuth, ocUser, ocCurrentCart, ocUnsubmittedOrders, hasCheckedForToken]);
+  }, [dispatch, ocAuth, ocUser, ocCurrentCart, hasCheckedForToken]);
 
   return <>{children}</>;
 };
