+++ Website/src/rendering/src/redux/store.ts
@@ -7,7 +7,6 @@ import ocProductCache from './ocProductCache';
 import ocPreviewSearch from './ocPreviewSearch';
 import ocProductList from './ocProductList';
 import ocProductDetail from './ocProductDetail';
-import ocUnsubmittedOrders from './ocUnsubmittedOrders';
 import ocCurrentCart from './ocCurrentCart';
 import ocAddressBook from './ocAddressBook';
 
@@ -22,7 +21,6 @@ const store = configureStore({
     ocProductList,
     ocProductDetail,
     ocCurrentCart,
-    ocUnsubmittedOrders,
   },
 });
